Split helper script tests into separate files
